Create compatibility link from cvd create
This links are meant to provide backwards compatibility with acloud and
launch_cvd, which only makes sense when starting the device at creation
time. It's also only at creation time that the check for default group
can be made as by the time start executes a default group has been
chosen that can't be differentiated from a custom one.
